--- keewe-domain/src/main/java/ccc/keewedomain/common/Photo.java
@@ -1,5 +1,0 @@
-package ccc.keewedomain.common;
-
-
-public class Photo {
-}
