+++ keewe-domain/src/main/java/ccc/keewedomain/domain/common/Photo.java
@@ -1,0 +1,5 @@
+package ccc.keewedomain.domain.common;
+
+
+public class Photo {
+}
