+++ keewe-domain/src/main/java/ccc/keewedomain/domain/user/ProfileLink.java
@@ -1,7 +1,8 @@
-package ccc.keewedomain.user;
+package ccc.keewedomain.domain.user;
 
 import ccc.keewedomain.common.BaseTimeEntity;
 import ccc.keewedomain.common.Link;
+
 import javax.persistence.*;
 
 import static javax.persistence.FetchType.LAZY;
