+++ keewe-domain/src/main/java/ccc/keewedomain/domain/user/enums/Privacy.java
@@ -1,4 +1,4 @@
-package ccc.keewedomain.user.enums;
+package ccc.keewedomain.domain.user.enums;
 
 public enum Privacy {
     PUBLIC,
